+++ frontend/src/pages/projects/[projectId]/reporting/index.vue
@@ -1,3 +1,20 @@
 <template>
-  <div></div>
+  <div class="w-100 d-flex flex-column text-center">
+    <img src="~/assets/dino/project.svg" alt="" class="img-raptor" />
+    <div>
+      <strong>Start writing report sections and findings</strong>
+      <p>
+        Select a section or finding from the menu on the left to start editing or create a new finding by clicking the add button. 
+      </p>
+    </div>
+  </div>
 </template>
+
+<style lang="scss" scoped>
+.img-raptor {
+  width: 30em;
+  max-width: 50%;
+  max-height: 50vh;
+  margin-top: 20vh;
+}
+</style>