+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix error while editing ID of nested field of report section in designer
 * Add demo data archives as TOML files to repository
 * Fix resizing PDF viewer loses mouse focus in Firefox
+* Add raptor mascot images as to empty pages
 
 
 ## v2024.3 - 2024-01-09